preparing to nicely animate dots as they are selected
Changed the class marked to selected, and colored these selected dots black
for visibility. Also selected the dots one at a time as a segment is drawn
through them, and marking all of the same colored dots at the moment a cycle
is completed, rather than all at once at the beginning.