vendorized animations and transitions, and fixed the square-completing logic
When completing a cycle that goes through a dot twice (e.g. a figure eight)
the frontend completed the cycle when it should have waited until the path
was fully completed. Also there was an overflow error when all 36 dots are the
same color (all of the cycles are found). Until a better solution to that is found,
I raised the size of the list of moves.